refactor: move Single and MultiSelect to rooted component folder
