--- react/src/Dropdown/SingleSelect/index.ts
@@ -1,1 +1,0 @@
-export * from './SingleSelect'
