--- react/src/Dropdown/components/DropdownItem/index.ts
@@ -1,1 +1,0 @@
-export * from './DropdownItem'
