+++ react/src/MultiSelect/components/MultiDropdownItem/ItemCheckboxIcon.tsx
