+++ react/src/MultiSelect/components/MultiDropdownItem/index.ts
