+++ react/src/MultiSelect/components/MultiSelectItem/index.ts
