+++ react/src/SingleSelect/components/SelectCombobox/ComboboxClearButton.tsx
