+++ react/src/SingleSelect/index.ts
@@ -1,3 +1,3 @@
-export * from './MultiSelect'
+export * from './SelectContext'
 export * from './SingleSelect'
 export * from './types'
