+++ react/src/SingleSelect/utils/defaultFilter.ts
