+++ react/src/index.ts
@@ -8,7 +8,6 @@ export * from './Checkbox'
 export * from './DatePicker'
 export * from './DateRangePicker'
 export * from './Drawer'
-export * from './Dropdown'
 export * from './FormControl'
 export * from './hooks'
 export * from './IconButton'
@@ -18,6 +17,7 @@ export * from './Input'
 export * from './Link'
 export * from './Menu'
 export * from './Modal'
+export * from './MultiSelect'
 export * from './NumberInput'
 export * from './Pagination'
 export * from './PhoneNumberInput'
@@ -26,6 +26,7 @@ export * from './Radio'
 export * from './RestrictedFooter'
 export * from './RestrictedGovtMasthead'
 export * from './Searchbar'
+export * from './SingleSelect'
 export * from './Spinner'
 export * from './Switch'
 export * from './Tabs'
